feat: re-work optimizations of PG and add support for pgbouncer
Instead of allowing massively large number of PG connections on larger
instances, we limit them to ~500, instead opting to allow greater
amounts of work_mem per connection. In concert with this change, we
also optimize PGBouncer and allow it to support increasing number of
users.

The optimizations for pgbouncer are based off resources[1] which
demonstrate not much value in the backing pool size being
significantly larger than 1-2x the number of CPU cores available.

The numbers for postgresql itself are based on a combination of a few
things:
- pgtune recommendations
- number of connections inspired by Heroku
- setting aside some amount of memory for pgbouncer, since it's
collocated